@@ -8,16 +8,17 @@
     "build": "webpack -p",
     "startxx": "webpack-dev-server -dependencies --open | bunyan",
     "migrate": "sequelize db:migrate | bunyan",
-    "server-start": "nodemon server.js | bunyan --output short --color",
+    "server-start": "nodemon -r esm server.js | bunyan --output short --color",
     "client-start": "webpack -d --watch",
     "start": "concurrently --kill-others \"npm run server-start\" \"npm run client-start\""
   },
   "dependencies": {
-    "@material-ui/core": "^1.2.0",
-    "@material-ui/icons": "^1.1.0",
-    "axios": "^0.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-preset-es2015": "^6.24.1",
+    "@babel/cli": "^7.6.0",
+    "@date-io/moment": "^1.3.11",
+    "@material-ui/core": "^4.4.2",
+    "@material-ui/icons": "^4.4.1",
+    "axios": "^0.18.1",
+    "babel-plugin-react-css-modules": "^5.2.6",
     "bcrypt-nodejs": "0.0.3",
     "bluebird": "^3.5.1",
     "body-parser": "^1.18.3",
@@ -25,46 +26,52 @@
     "classnames": "^2.2.5",
     "construx": "^1.0.1",
     "construx-copier": "^1.0.0",
+    "core-js": "^3.2.1",
     "eslint": "^4.19.1",
+    "esm": "^3.2.25",
     "express": "^4.16.3",
     "helmet": "^3.12.1",
+    "history": "^4.10.1",
     "jsonwebtoken": "^8.3.0",
     "kraken-js": "^2.2.0",
-    "lodash": "^4.17.10",
-    "material-ui-pickers": "^1.0.0-rc.12",
-    "moment": "^2.22.2",
+    "lodash": "^4.17.15",
+    "moment": "^2.24.0",
     "moment-range": "^4.0.1",
-    "mongoose": "^5.1.4",
-    "morgan": "^1.9.0",
+    "mongoose": "^5.7.1",
+    "morgan": "^1.9.1",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-grid-system": "^5.0.2",
     "react-helmet": "^5.2.0",
     "react-moment": "^0.7.0",
     "react-router-dom": "^4.2.2",
+    "regenerator-runtime": "^0.13.3",
     "reset-css": "^4.0.1"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
+    "@babel/core": "^7.6.0",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/preset-env": "^7.6.0",
+    "@babel/preset-react": "^7.0.0",
+    "babel-loader": "^8.0.6",
     "concurrently": "^3.5.1",
-    "css-loader": "^0.28.11",
+    "css-loader": "^3.2.0",
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.2.0",
-    "img-loader": "^3.0.0",
+    "imagemin": "^6.0.0",
+    "img-loader": "^3.0.1",
+    "material-ui-pickers": "^2.2.4",
     "mocha": "^5.2.0",
-    "node-sass": "^4.9.0",
+    "node-sass": "^4.12.0",
     "nodemon": "^1.17.5",
     "sass-loader": "^7.0.3",
     "style-loader": "^0.21.0",
     "supertest": "^3.1.0",
     "url-loader": "^1.0.1",
-    "webpack": "^3.12.0",
-    "webpack-cli": "^2.1.5",
-    "webpack-dev-server": "^3.1.4"
+    "webpack": "^4.40.2",
+    "webpack-cli": "^3.3.8",
+    "webpack-dev-server": "^3.8.0"
   }
 }
