@@ -1,12 +1,16 @@
-var mongoose = require('mongoose');
+import mongoose from 'mongoose'
+import Division from './Division'
+import Club from './Club'
 
 var teamSchema = mongoose.Schema({
-  labelLocal: String,
+  labelClub: String,
+  labelClubShort: String,
+  labelDivision: String,
   short: String,
   key: String,
   _old: Number,
   label: String,
-  club: Object,
+  club: { type: mongoose.Schema.Types.ObjectId, ref: 'Club' },
   prefix: String,
   division: { type: mongoose.Schema.Types.ObjectId, ref: 'Division' },
   matchesWon: { type: Number, default: 0 },
@@ -25,4 +29,28 @@ var teamSchema = mongoose.Schema({
   pts: { type: Number, default: 0 }
 });
 
-module.exports = mongoose.model('Team', teamSchema);
+teamSchema.pre('save', function (next) {
+  var team = this;
+  Division.findOne({_id: team.division}).exec((err, division) => {
+    if (division) {
+      let lbl = division.category
+      if (team.prefix.length>0) lbl += ' ' + team.prefix
+      team.key = (lbl + '-' + division.key).toLowerCase().split(' ').join('-')
+      team.labelDivision = lbl
+      Club.findOne({_id: team.club}).exec((err, club) => {
+        if (club) {
+          let lbl = club.name
+          if (team.prefix.length>0) lbl += ' ' + team.prefix
+          lbl = club.short
+          if (team.prefix.length>0) lbl += ' ' + team.prefix
+          team.labelClubShort = lbl
+          return next();
+        }
+      })
+    }
+  })
+});
+
+const Team = mongoose.model('Team', teamSchema)
+
+export default Team
