@@ -9,15 +9,12 @@ class Loader extends Component {
 
     const style = this.props.loading ? { opacity:0.3 } : {}
 
-    // return (
-    //   <div className="loading animated fadeIn">
-    //     <div className="bg" style={style}></div>
-    //   </div>
-    // );
-
     return (
-      <span>loading: {this.props.loading ? 'true' : 'false'}</span>
+      <div className="loading animated fadeIn">
+        <div className="bg" style={style}></div>
+      </div>
     );
+
   }
 }
 
